+++ autologger/autologger.go
@@ -9,9 +9,6 @@ import (
 
 	log "log/slog"
 
-	"github.com/Azure/aks-middleware/common"
-	"github.com/Azure/aks-middleware/requestid"
-
 	"github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/logging"
 )
 
@@ -45,16 +42,4 @@ func InterceptorLogger(logger *log.Logger) logging.Logger {
 			panic(fmt.Sprintf("unknown level %v", lvl))
 		}
 	})
-}
-
-// Add request-id & headers to API autologger.
-func GetFields(ctx context.Context) logging.Fields {
-	headers := requestid.GetMetadata(ctx)
-	requestID := headers[common.RequestIDMetadataKey]
-	// Remove the main request ID from headers map since it's logged separately
-	delete(headers, common.RequestIDMetadataKey)
-	return logging.Fields{
-		common.RequestIDLogKey, requestID,
-		"headers", headers,
-	}
 }
